Update PdfProcess.cs
Found a couple of issues with the scale and rotate that are now fixed. If you pass in 0x and 0y, it will look at the existing page, figure out the rotation it is supposed to have, and bring it in line. The output should be all the pages the same way and 0 rotation applied. if you run a pdfinfo, all the pages should be the same way with 0 rotation applied, but the content should be rotated the correct way (via matrix).